backend/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
Werkzeug==3.1.3

pymupdf==1.26.4
pypdf==6.0.0
pdfplumber==0.11.7

docx2txt==0.9
python-docx==1.2.0

markdown==3.8.2
beautifulsoup4==4.13.4
chardet==5.2.0

pytest==8.4.1
ruff==0.12.8
requests==2.32.3
python-dotenv==1.0.1
pytesseract==0.3.13
